sdk!: Remove re-exports
#### Problem

It's time to make breaking changes in the sdk! There are a lot of
deprecated symbols all over the repository, especially in solana-sdk.

#### Summary of changes

Remove all deprecated symbols and dependencies from solana-sdk, and
include upgrade information in the README for people porting their code.

BREAKING CHANGE: many re-exports and symbols removed, documented in the
README