rent!: Update default values, remove deprecated code
#### Problem

Rent is gone, and SIMD-0194 will deprecate the exemption threshold, but
the field still exists in the actual sysvar on-chain.

#### Summary of changes

Remove all deprecated code related to collecting rent, and use the new
default values for Rent after SIMD-0194.

To go with this, avoid floating point math in `minimum_balance()`, which
was one of the of motivations for the change in the first place.

Add a test to make sure the old and new calcs are the same.